src/main/java/com/crm/qa/pages/ContactsPage.java
package com.crm.qa.pages;

import java.util.Iterator;
import java.util.List;

import org.openqa.selenium.By;
import org.openqa.selenium.WebElement;
import org.openqa.selenium.support.FindBy;
import org.openqa.selenium.support.PageFactory;
import org.openqa.selenium.support.ui.Select;

import com.crm.qa.base.TestBase;

public class ContactsPage extends TestBase {
	
	
	@FindBy(xpath = "//td[contains(text(),'Contacts')]")
	WebElement contactsLabel;
	
	@FindBy(name = "first_name")
	WebElement firstName;
	
	@FindBy(name="surname")
	WebElement lastName;
	
	@FindBy(name = "client_lookup")
	WebElement company;
	
	@FindBy(xpath = "//input[@type='submit' and @value='Save']")
	WebElement saveContact;
	
	
	public ContactsPage() {
		PageFactory.initElements(driver, this);
	}
	
//	@FindBy(xpath = "//a[contains(text(),'David Cris')]//parent::td[@class='datalistrow']//preceding-sibling::td[@class='datalistrow']/input[@name='contact_id']")
//	WebElement 
	
	public boolean verifyContactsLabel() {
		return contactsLabel.isDisplayed();
	}
	
	public boolean selectContactsByName(String name) {
		driver.findElement(By.xpath("//a[contains(text(),'"+name+"')]//parent::td[@class='datalistrow']//preceding-sibling::td[@class='datalistrow']/input[@name='contact_id']")).click();
		boolean b =driver.findElement(By.xpath("//a[contains(text(),'"+name+"')]//parent::td[@class='datalistrow']//preceding-sibling::td[@class='datalistrow']/input[@name='contact_id']")).isSelected();
		return b;
	}
	
	public void createNewContact(String title,String ftname,String ltname,String comp) {
		Select sc= new Select(driver.findElement(By.name("title")));
		sc.selectByVisibleText(title);
		
		firstName.sendKeys(ftname);
		lastName.sendKeys(ltname);
		company.sendKeys(comp);
		
		saveContact.click();
	}
	
		
	}
